move messageConfig to its own folder
